requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.30.0
google-cloud-translate>=3.15.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
moviepy>=1.0.3
elevenlabs>=1.0.0
lxml>=5.0.0
Werkzeug>=3.0.0
gunicorn>=21.0.0
